refactor: validate in service not state
It is not the concern of the state code to validate what the input is,
that is the concern of the service code. Otherwise what's the point of
the service?
