+++ domain/crossmodelrelation/service/remoteapplication.go
@@ -9,6 +9,8 @@ import (
 
 	"gopkg.in/macaroon.v2"
 
+	"github.com/juju/collections/set"
+	"github.com/juju/collections/transform"
 	coreapplication "github.com/juju/juju/core/application"
 	corecharm "github.com/juju/juju/core/charm"
 	"github.com/juju/juju/core/errors"
@@ -273,6 +275,13 @@ func (s *Service) AddConsumedRelation(ctx context.Context, args AddConsumedRelat
 	charmUUID, err := corecharm.NewID()
 	if err != nil {
 		return internalerrors.Errorf("creating charm uuid: %w", err)
+	}
+
+	// Check that the charm has only one endpoint. There can be multiple
+	// synthetic applications per offer, but only one endpoint per synthetic
+	// application. To do otherwise requires design and facade changes.
+	if err := synthCharmHasOnlyOneEndpoint(args.ConsumerApplicationEndpoint.Name, syntheticCharm); err != nil {
+		return internalerrors.Errorf("adding consumed relation: %w", err)
 	}
 
 	if err := s.modelState.AddConsumedRelation(ctx, synthApplicationName, crossmodelrelation.AddRemoteApplicationConsumerArgs{
@@ -627,4 +636,22 @@ func (s *Service) GetRelationRemoteModelUUID(ctx context.Context, relationUUID c
 		return coremodel.UUID(""), internalerrors.Capture(err)
 	}
 	return modelUUID, nil
+}
+
+func synthCharmHasOnlyOneEndpoint(endpoint string, ch charm.Charm) error {
+	provides := transform.MapToSlice(ch.Metadata.Provides, func(k string, _ charm.Relation) []string {
+		return []string{k}
+	})
+	requires := transform.MapToSlice(ch.Metadata.Requires, func(k string, _ charm.Relation) []string {
+		return []string{k}
+	})
+	providesSet := set.NewStrings(provides...)
+	requiresSet := set.NewStrings(requires...)
+	cnt := providesSet.Size() + requiresSet.Size()
+	if providesSet.Union(requiresSet).Contains(endpoint) && cnt == 1 {
+		return nil
+	} else if cnt > 1 {
+		return internalerrors.Errorf("application in relation has more than one potential endpoint").Add(relationerrors.AmbiguousRelation)
+	}
+	return internalerrors.Errorf("endpoint %q", endpoint).Add(relationerrors.RelationEndpointNotFound)
 }
